+++ CHANGELOG.md
@@ -12,7 +12,7 @@ For details about compatibility between different releases, see the **Commitment
 ### Added
 
 - Allow setting an expiry date for API keys in the Console
-- New event type `gs.gateway.connection.stats` with connection statistics. A new event is sent at most every `gs.update-connection-stats-debounce-time` time.
+- New event type `gs.gateway.connection.stats` with connection statistics. A new event is sent at most every `gs.update-connection-stats-debounce-time` time and at least every `gs.update-connection-stats-interval` time.
 - Button to export as JSON end device `mac-settings` and `mac-state` in the Console.
 
 ### Changed